SiamesePytorch/requirements.txt
scipy
numpy
matplotlib
opencv_python
# torch==1.10.0
# torchvision==0.11.0
tqdm
Pillow
h5py
